+++ Sources/Views/Cells/TextMessageCell.swift
@@ -24,64 +24,77 @@
 
 import UIKit
 
-open class TextMessageCell: MessageCollectionViewCell<MessageLabel> {
+open class TextMessageCell: MessageCollectionViewCell {
+
     open override class func reuseIdentifier() -> String { return "messagekit.cell.text" }
 
     // MARK: - Properties
 
     open override weak var delegate: MessageCellDelegate? {
         didSet {
-            messageContentView.delegate = delegate
+            messageLabel.delegate = delegate
         }
     }
 
-    override var messageTapGesture: UITapGestureRecognizer? {
-        didSet {
-            messageTapGesture?.delegate = messageContentView
-        }
-    }
+    open var messageLabel = MessageLabel()
 
     // MARK: - Methods
 
     open override func apply(_ layoutAttributes: UICollectionViewLayoutAttributes) {
         super.apply(layoutAttributes)
-
-        guard let attributes = layoutAttributes as? MessagesCollectionViewLayoutAttributes else { return }
-        
-        messageContentView.configure {
-            messageContentView.textInsets = attributes.messageLabelInsets
-            messageContentView.font = attributes.messageLabelFont
+        if let attributes = layoutAttributes as? MessagesCollectionViewLayoutAttributes {
+            messageLabel.textInsets = attributes.messageLabelInsets
+            messageLabel.font = attributes.messageLabelFont
         }
-
     }
 
     open override func prepareForReuse() {
         super.prepareForReuse()
-        messageContentView.attributedText = nil
-        messageContentView.text = nil
+        messageLabel.attributedText = nil
+        messageLabel.text = nil
+    }
+
+    open override func setupSubviews() {
+        super.setupSubviews()
+        messageContainerView.addSubview(messageLabel)
+        setupConstraints()
+    }
+    
+    open func setupConstraints() {
+        messageLabel.fillSuperview()
     }
 
     open override func configure(with message: MessageType, at indexPath: IndexPath, and messagesCollectionView: MessagesCollectionView) {
         super.configure(with: message, at: indexPath, and: messagesCollectionView)
 
-        if let displayDelegate = messagesCollectionView.messagesDisplayDelegate {
-            let textColor = displayDelegate.textColor(for: message, at: indexPath, in: messagesCollectionView)
-            let detectors = displayDelegate.enabledDetectors(for: message, at: indexPath, in: messagesCollectionView)
-            
-            messageContentView.configure {
-                messageContentView.textColor = textColor
-                messageContentView.enabledDetectors = detectors
-            }
+        guard let displayDelegate = messagesCollectionView.messagesDisplayDelegate else {
+            fatalError("MessagesDisplayDelegate not set.")
         }
 
-        switch message.data {
-        case .text(let text), .emoji(let text):
-            messageContentView.text = text
-        case .attributedText(let text):
-            messageContentView.attributedText = text
-        default:
-            break
+        let textColor = displayDelegate.textColor(for: message, at: indexPath, in: messagesCollectionView)
+        let enabledDetectors = displayDelegate.enabledDetectors(for: message, at: indexPath, in: messagesCollectionView)
+
+        messageLabel.configure {
+            messageLabel.enabledDetectors = enabledDetectors
+            for detector in enabledDetectors {
+                let attributes = displayDelegate.detectorAttributes(for: detector, and: message, at: indexPath)
+                messageLabel.setAttributes(attributes, detector: detector)
+            }
+            switch message.data {
+            case .text(let text), .emoji(let text):
+                messageLabel.text = text
+            case .attributedText(let text):
+                messageLabel.attributedText = text
+            default:
+                break
+            }
+            // Needs to be set after the attributedText because it takes precedence
+            messageLabel.textColor = textColor
         }
     }
     
+    /// Handle `ContentView`'s tap gesture, return false when `ContentView` don't needs to handle gesture
+    open override func cellContentView(canHandle touchPoint: CGPoint) -> Bool {
+        return messageLabel.handleGesture(touchPoint)
+    }
 }