+++ Tests/SenderSpec.swift
@@ -1,18 +1,18 @@
 /*
  MIT License
- 
+
  Copyright (c) 2017 MessageKit
- 
+
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
- 
+
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
- 
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -22,15 +22,28 @@
  SOFTWARE.
  */
 
-import XCTest
+import Quick
+import Nimble
 @testable import MessageKit
 
-class AvatarTests: XCTestCase {
+final class SenderSpec: QuickSpec {
 
-    func testDefaultInit() {
-        let avatar = Avatar()
-        XCTAssertNil(avatar.image)
-        XCTAssertEqual(avatar.initials, "?")
+    override func spec() {
+        describe("equality between two Senders") {
+            context("they have the same id ") {
+                it("should be equal") {
+                    let sender1 = Sender(id: "1", displayName: "Steven")
+                    let sender2 = Sender(id: "1", displayName: "Nathan")
+                    expect(sender1).to(equal(sender2))
+                }
+            }
+            context("they have a different id") {
+                it("should not be equal") {
+                    let sender1 = Sender(id: "1", displayName: "Steven")
+                    let sender2 = Sender(id: "2", displayName: "Nathan")
+                    expect(sender1).toNot(equal(sender2))
+                }
+            }
+        }
     }
-
 }